+++ flobank-api/src/main/java/kr/co/api/flobankapi/mapper/CustAcctMapper.java
@@ -1,5 +1,8 @@
 package kr.co.api.flobankapi.mapper;
 
+import java.math.BigDecimal;
+
 public interface CustAcctMapper {
-    void updateKoAcctBal(Integer bal, String acctNo);
+    // 계좌 출금되었을 때 금액 업데이트
+    void updateKoAcctBal(BigDecimal bal, String acctNo);
 }
